requirements.txt
pydantic>=2.7.0
python-dotenv>=1.0.0
openai>=1.30.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
